+++ backend/CHANGELOG.md
@@ -1,3 +1,13 @@
+# [backend-v1.1.0](https://github.com/Darchie4/DevOps-Hand-in-G09a/compare/backend-v1.0.0...backend-v1.1.0) (2023-08-20)
+
+
+### Features
+
+* **backend:** added simple response to "/healtz" ([7487db4](https://github.com/Darchie4/DevOps-Hand-in-G09a/commit/7487db4445843baf04cd57e5d86e81588a1f9f69))
+* **backend-test:** added simple test to /healtz ([ae1c3e4](https://github.com/Darchie4/DevOps-Hand-in-G09a/commit/ae1c3e4339f95e9a6f474d09dfc99db3d70e6d4c))
+* **backend-test:** finished adding /healthz test to backend ([a1cf6e0](https://github.com/Darchie4/DevOps-Hand-in-G09a/commit/a1cf6e0122be85865a4c27c96cbc555af3f9b1c5))
+* **backend-test:** finished work on /healtz test ([b461392](https://github.com/Darchie4/DevOps-Hand-in-G09a/commit/b4613920df10d457dffcb7f3406e761c610e2abb))
+
 # backend-v1.0.0 (2023-08-19)